# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuruForge API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  icon?: string | null;
  displayOrder: number;
}

export interface GuruListItem {
  id: number;
  name: string;
  slug: string;
  tagline?: string | null;
  description?: string | null;
  categoryId?: number | null;
  avatarUrl?: string | null;
  status: string;
  priceCents: number;
  priceInterval: string;
  topics?: string[] | null;
  personalityStyle?: string | null;
  modelTier?: string | null;
  wisdomScore?: number | null;
  satisfactionScore?: number | null;
  userCount?: number | null;
  createdAt: string;
  categoryName?: string | null;
  categorySlug?: string | null;
  creatorName?: string | null;
}

export interface GuruDetail {
  id: number;
  creatorId: number;
  name: string;
  slug: string;
  tagline?: string | null;
  description?: string | null;
  categoryId?: number | null;
  avatarUrl?: string | null;
  status: string;
  priceCents: number;
  priceInterval: string;
  topics?: string[] | null;
  personalityStyle?: string | null;
  modelTier?: string | null;
  memoryPolicy?: string | null;
  introEnabled?: boolean | null;
  wisdomScore?: number | null;
  satisfactionScore?: number | null;
  userCount?: number | null;
  createdAt: string;
  categoryName?: string | null;
  categorySlug?: string | null;
  creatorName?: string | null;
  creatorAvatarUrl?: string | null;
  avgRating?: number | null;
  totalRatings: number;
}

export interface Guru {
  id: number;
  creatorId: number;
  name: string;
  slug: string;
  tagline?: string | null;
  description?: string | null;
  categoryId?: number | null;
  avatarUrl?: string | null;
  status: string;
  priceCents: number;
  priceInterval: string;
  topics?: string[] | null;
  personalityStyle?: string | null;
  modelTier?: string | null;
  memoryPolicy?: string | null;
  introEnabled?: boolean | null;
  wisdomScore?: number | null;
  satisfactionScore?: number | null;
  userCount?: number | null;
  stripeProductId?: string | null;
  stripePriceId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateGuruInputPriceInterval =
  (typeof CreateGuruInputPriceInterval)[keyof typeof CreateGuruInputPriceInterval];

export const CreateGuruInputPriceInterval = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

export type CreateGuruInputPersonalityStyle =
  (typeof CreateGuruInputPersonalityStyle)[keyof typeof CreateGuruInputPersonalityStyle];

export const CreateGuruInputPersonalityStyle = {
  professional: "professional",
  friendly: "friendly",
  direct: "direct",
  academic: "academic",
} as const;

export type CreateGuruInputModelTier =
  (typeof CreateGuruInputModelTier)[keyof typeof CreateGuruInputModelTier];

export const CreateGuruInputModelTier = {
  basic: "basic",
  pro: "pro",
  enterprise: "enterprise",
} as const;

export interface CreateGuruInput {
  name: string;
  tagline?: string;
  description?: string;
  categoryId?: number;
  avatarUrl?: string;
  priceCents?: number;
  priceInterval?: CreateGuruInputPriceInterval;
  topics?: string[];
  personalityStyle?: CreateGuruInputPersonalityStyle;
  modelTier?: CreateGuruInputModelTier;
  memoryPolicy?: string;
  introEnabled?: boolean;
}

export type UpdateGuruInputStatus =
  (typeof UpdateGuruInputStatus)[keyof typeof UpdateGuruInputStatus];

export const UpdateGuruInputStatus = {
  draft: "draft",
  published: "published",
  archived: "archived",
} as const;

export type UpdateGuruInputPriceInterval =
  (typeof UpdateGuruInputPriceInterval)[keyof typeof UpdateGuruInputPriceInterval];

export const UpdateGuruInputPriceInterval = {
  monthly: "monthly",
  yearly: "yearly",
} as const;

export type UpdateGuruInputPersonalityStyle =
  (typeof UpdateGuruInputPersonalityStyle)[keyof typeof UpdateGuruInputPersonalityStyle];

export const UpdateGuruInputPersonalityStyle = {
  professional: "professional",
  friendly: "friendly",
  direct: "direct",
  academic: "academic",
} as const;

export type UpdateGuruInputModelTier =
  (typeof UpdateGuruInputModelTier)[keyof typeof UpdateGuruInputModelTier];

export const UpdateGuruInputModelTier = {
  basic: "basic",
  pro: "pro",
  enterprise: "enterprise",
} as const;

export interface UpdateGuruInput {
  name?: string;
  tagline?: string;
  description?: string;
  categoryId?: number;
  avatarUrl?: string;
  status?: UpdateGuruInputStatus;
  priceCents?: number;
  priceInterval?: UpdateGuruInputPriceInterval;
  topics?: string[];
  personalityStyle?: UpdateGuruInputPersonalityStyle;
  modelTier?: UpdateGuruInputModelTier;
  memoryPolicy?: string;
  introEnabled?: boolean;
  telegramBotToken?: string;
}

export interface User {
  id: number;
  clerkId: string;
  email: string;
  name?: string | null;
  avatarUrl?: string | null;
  role: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateUserInput {
  name?: string;
  avatarUrl?: string;
}

export interface Rating {
  id: number;
  rating: number;
  comment?: string | null;
  createdAt: string;
  userName?: string | null;
  userAvatarUrl?: string | null;
}

export interface GuruRating {
  id: number;
  userId: number;
  guruId: number;
  rating: number;
  comment?: string | null;
  createdAt: string;
}

export interface CreateRatingInput {
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string;
}

export interface CheckoutInput {
  guruId: number;
}

export interface CheckoutSession {
  url: string | null;
}

export interface UserSubscription {
  id: number;
  guruId: number;
  status: string;
  startedAt: string;
  expiresAt?: string | null;
  stripeSubscriptionId?: string | null;
  guruName: string;
  guruSlug: string;
  guruAvatarUrl?: string | null;
  guruPriceCents: number;
  guruPriceInterval: string;
}

export interface PortalSession {
  url: string;
}

export interface SubscriptionCheck {
  subscribed: boolean;
}

export interface TelegramConnectionResponse {
  code?: string | null;
  expiresAt?: string | null;
  expiresInSeconds?: number | null;
  alreadyConnected?: boolean | null;
  message?: string | null;
  botUsername?: string | null;
  botLink?: string | null;
}

export interface TelegramStatusResponse {
  connected: boolean;
  connectedAt?: string | null;
  contributesToWisdom?: boolean;
}

export interface TelegramBotInfoResponse {
  configured: boolean;
  botUsername?: string | null;
}

export interface WisdomToggleInput {
  contributesToWisdom: boolean;
}

export interface WisdomToggleResponse {
  contributesToWisdom: boolean;
}

export interface ContributionScoreResponse {
  score: number;
  conversationCount: number;
  patternsContributed: number;
  lastUpdatedAt?: string | null;
}

export interface UpdateBotTokenInput {
  botToken: string;
}

export interface UpdateBotTokenResponse {
  success: boolean;
  botUsername?: string | null;
  botName?: string | null;
}

export type ListGurusParams = {
  /**
   * Filter by category slug
   */
  category?: string;
  /**
   * Filter by status (default published)
   */
  status?: string;
  /**
   * Search by name, tagline, or description
   */
  search?: string;
  /**
   * Sort order
   */
  sort?: ListGurusSort;
};

export type ListGurusSort = (typeof ListGurusSort)[keyof typeof ListGurusSort];

export const ListGurusSort = {
  newest: "newest",
  wisdom: "wisdom",
  price_asc: "price_asc",
  price_desc: "price_desc",
  users: "users",
} as const;