# lib/api-zod/src/generated/types/contributionScoreResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuruForge API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContributionScoreResponse {
  score: number;
  conversationCount: number;
  patternsContributed: number;
  lastUpdatedAt?: Date | null;
}